requirements.txt
asyncpg
redis
pydantic
discord.py
python-dotenv
colorlog
pytz